+++ NEWS.md
@@ -1,3 +1,11 @@
 # SPA 0.0.0
 
 * Initial CRAN submission.
+
+# SPA 0.1.0
+
+* Apply SPA method in logistic regression model.
+
+# SPA 0.2.0
+
+* Apply SPA method in NB generalized linear model.